Realizado alterações nas estruturas das entidades, e finalizado crud de pessoa e empresa.

+++ LixoEletronico.Domain/Contracts/IPersonRepository.cs
@@ -4,8 +4,8 @@ namespace LixoEletronico.Domain.Contracts
 {
     public interface IPersonRepository
     {
-        Task UpdatePerson(long id, Person person);
+        Task UpdatePerson(int id, Person person);
 
-        Task<Person> GetPerson(long id);
+        Task<Person> GetPerson(int id);
     }
 }
